comments, explanations, fixes
